refactored addon dummy apps
