+++ packages/codegen/tests/dummy/cardstack/initial-data/data-sources/ephemeral.js
