+++ packages/github-auth/tests/dummy/cardstack/initial-data/data-sources/ephemeral.js
