+++ packages/hub/index.js
@@ -45,7 +45,7 @@ let addon = {
   async url() {
     if (!this._hub) {
       this._env = process.env.EMBER_ENV || 'development';
-      if (fs.existsSync(path.join(path.dirname(this.project.configPath()), '..', 'cardstack', 'seeds', this._env))) {
+      if (fs.existsSync(path.join(path.dirname(this.project.configPath()), '..', 'cardstack', 'initial-data'))) {
         this._hub = this._startHub();
       } else {
         this._hub = Promise.resolve(null);
