+++ packages/live-queries/tests/dummy/cardstack/initial-data/seeds/ephemeral.js
@@ -1,0 +1,17 @@
+/* eslint-env node */
+const JSONAPIFactory = require('@cardstack/test-support/jsonapi-factory');
+
+function initialModels() {
+  let factory = new JSONAPIFactory();
+
+  /*
+  factory.addResource('items')
+    .withAttributes({
+      content: 'hello'
+    });
+  */
+
+  return factory.getModels()
+}
+
+module.exports = initialModels();
