+++ packages/mock-auth/tests/dummy/cardstack/initial-data/data-sources/ephemeral.js
@@ -1,0 +1,21 @@
+/* eslint-env node */
+
+module.exports = [
+  {
+    type: 'data-sources',
+    id: 'default',
+    attributes: {
+      'source-type': '@cardstack/ephemeral',
+    }
+  },
+  {
+    type: 'plugin-configs',
+    id: '@cardstack/hub',
+    relationships: {
+      'default-data-source': {
+        data: { type: 'data-sources', id: 'default' }
+      }
+    }
+  }
+];
+
