--- packages/models/tests/dummy/cardstack/seeds/development/ephemeral.js
@@ -1,44 +1,0 @@
-/* eslint-env node */
-const JSONAPIFactory = require('@cardstack/test-support/jsonapi-factory');
-
-function initialModels() {
-  let initial = new JSONAPIFactory();
-
-  initial.addResource('content-types', 'posts')
-    .withRelated('fields', [
-      initial.addResource('fields', 'title').withAttributes({
-        fieldType: '@cardstack/core-types::string'
-      })
-    ]);
-  initial.addResource('posts', '1')
-    .withAttributes({
-      title: 'hello world'
-    });
-  initial.addResource('posts', '2')
-    .withAttributes({
-      title: 'second'
-    });
-  return initial.getModels();
-}
-
-module.exports = [
-  {
-    type: 'data-sources',
-    id: 'default',
-    attributes: {
-      'source-type': '@cardstack/ephemeral',
-      params: {
-        initialModels: initialModels()
-      }
-    }
-  },
-  {
-    type: 'plugin-configs',
-    id: '@cardstack/hub',
-    relationships: {
-      'default-data-source': {
-        data: { type: 'data-sources', id: 'default' }
-      }
-    }
-  }
-];
