--- packages/routing/tests/dummy/cardstack/seeds/development/ephemeral.js
@@ -1,26 +1,0 @@
-/* eslint-env node */
-
-const { initialModels } = require('../test/ephemeral');
-
-
-module.exports = [
-  {
-    type: 'data-sources',
-    id: 'default',
-    attributes: {
-      'source-type': '@cardstack/ephemeral',
-      params: {
-        initialModels: initialModels()
-      }
-    }
-  },
-  {
-    type: 'plugin-configs',
-    id: '@cardstack/hub',
-    relationships: {
-      'default-data-source': {
-        data: { type: 'data-sources', id: 'default' }
-      }
-    }
-  }
-];
