--- packages/tools/tests/dummy/cardstack/seeds/test/test.js
@@ -1,4 +1,0 @@
-/* eslint-env node */
-
-// Use the same seed models in test as in development
-module.exports = require('../development/dev');
